+++ examples/11_germanium_solid/11_germanium_solid.py
@@ -10,7 +10,7 @@ from eminus.units import ang2bohr, ev2ha
 atom = 'Ge8'
 a = ang2bohr(5.658)
 ecut = ev2ha(500)
-X = a * np.array([
+pos = a * np.array([
     [0, 0, 0],
     [0, 0.5, 0.5],
     [0.25, 0.25, 0.25],
@@ -20,7 +20,7 @@ X = a * np.array([
     [0.75, 0.25, 0.25],
     [0.75, 0.75, 0.25]
 ])
-atoms = Atoms(atom=atom, X=X, a=a, ecut=ecut)
+atoms = Atoms(atom=atom, pos=pos, a=a, ecut=ecut)
 
 # # Use the pseudopotential from Tomas Arias (only local and only for germanium)
 # # The GTH pseudopotential will work as well and would include non-local effects